Additional refactor; settings cleanup
