+++ app/src/main/java/vadimerenkov/aucards/MainActivity.kt
@@ -10,7 +10,7 @@ import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.lifecycle.lifecycleScope
-import vadimerenkov.aucards.settings.Theme
+import vadimerenkov.aucards.screens.settings.Theme
 import vadimerenkov.aucards.ui.theme.AucardsTheme
 
 class MainActivity : AppCompatActivity() {
