+++ app/src/main/java/vadimerenkov/aucards/screens/settings/Constants.kt
@@ -1,0 +1,7 @@
+package vadimerenkov.aucards.screens.settings
+
+const val THEME_STRING = "theme"
+const val BRIGHTNESS_STRING = "brightness"
+const val LANDSCAPE_STRING = "landscape"
+const val SOUND_STRING = "play_sound"
+const val RINGTONE_URI = "ringtone_uri"
