--- app/src/main/java/vadimerenkov/aucards/settings/Keys.kt
@@ -1,9 +1,0 @@
-package vadimerenkov.aucards.settings
-
-object Keys {
-	const val THEME_STRING = "theme"
-	const val BRIGHTNESS_STRING = "brightness"
-	const val LANDSCAPE_STRING = "landscape"
-	const val SOUND_STRING = "play_sound"
-	const val RINGTONE_URI = "ringtone_uri"
-}
